% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadRedcapReport.R
\name{downloadRedcapReport}
\alias{downloadRedcapReport}
\title{Title: Download REDCap report}
\usage{
downloadRedcapReport(redcapTokenName, redcapUrl, redcapReportId)
}
\arguments{
\item{redcapTokenName}{The name of the API token stored in the .REnviron file.}

\item{redcapUrl}{The URL of the REDCap API.}

\item{redcapReportId}{The ID of the REDCap report to retrieve.}
}
\value{
A tibble containing the contents of the REDCap report.
}
\description{
This function retrieves a REDCap report using the provided API token, URL, and report ID.
}
