% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikBernoulli.R
\name{logLikBernoulli}
\alias{logLikBernoulli}
\title{Title: Maximum Likelihood Estimate}
\usage{
logLikBernoulli(data)
}
\arguments{
\item{data}{A numeric vector of binary outcomes (0s and 1s).}
}
\value{
The maximum likelihood estimate of parameter p.
}
\description{
Given a vector of binary outcomes, calculates the parameter p that maximizes
the log-likelihood of the data under the assumption of a Bernoulli distribution.
}
