% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimumN.R
\name{minimumN}
\alias{minimumN}
\title{Title: Minimum sample size calculation for a t-test}
\usage{
minimumN(x1, x2 = NULL)
}
\arguments{
\item{x1}{A numeric vector representing the first sample.}

\item{x2}{A numeric vector representing the second sample (optional).}
}
\value{
The minimum sample size needed for a t-test.
}
\description{
This function is a wrapper around `pwr::pwr.t2n.test` that calculates the minimum sample size
needed for a t-test of the null hypothesis with specified power and significance level.
}
\examples{
set.seed(123)
sample_data1 <- rnorm(50)
sample_data2 <- rnorm(100)
minimumN(sample_data1, sample_data2)
#' @export
}
