% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeNames.R
\name{standardizeNames}
\alias{standardizeNames}
\title{Title: Standardize variable names}
\usage{
standardizeNames(data)
}
\arguments{
\item{data}{A tibble.}
}
\value{
The tibble with standardized variable names.
}
\description{
Title: Standardize variable names
}
