% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survCurv.R
\name{survCurv}
\alias{survCurv}
\title{Title: Survival curve}
\usage{
survCurv(status, time)
}
\arguments{
\item{status}{A numeric vector indicating event status (0 for censored, 1 for event).}

\item{time}{A numeric vector of time to event or censoring.}
}
\value{
A plot of the survival curve.
}
\description{
Given vectors of event status and corresponding times, calculates and plots
the survival curve S(t).
}
\examples{
# Example usage:
library(survival)
test_survCurv <- function() {
  url <- "https://jlucasmckay.bmi.emory.edu/global/bmi510/Labs-Materials/survival.csv"
  survival_data <- read.csv(url)
  str(survival_data)
  survCurv(status = survival_data$status, time = survival_data$time)
}
# Call the test function
test_survCurv()
}
