% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscale.R
\name{unscale}
\alias{unscale}
\title{Title: Reverse the centering/scaling of a vector}
\usage{
unscale(x)
}
\arguments{
\item{x}{A numeric vector that has been centered and scaled.}
}
\value{
The unscaled and uncentered vector.
}
\description{
Takes a vector that has been put through scaling and reverses the centering/scaling, if any.
}
\examples{
scaled_vector <- scale(c(1, 2, 3, 4, 5))
unscaled_vector <- unscale(scaled_vector)
}
