% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeNames.R
\name{standardizeNames}
\alias{standardizeNames}
\title{Title: Standardize variable names}
\usage{
standardizeNames(data)
}
\arguments{
\item{data}{A tibble.}
}
\value{
The tibble with standardized variable names.
}
\description{
Title: Standardize variable names
}
\examples{
# Example 1: Standardize variable names in a tibble
library(tibble)
data <- tibble("First Name" = c("Trey", "Lina"),
               "Last Name" = c("Travis", "Kaushik"),
               "Age" = c(22, 23))
clean_data <- standardizeNames(data)
print(clean_data)
}
