% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscale.R
\name{unscale}
\alias{unscale}
\title{Title: Reverse the centering/scaling of a vector}
\usage{
unscale(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The unscaled vector.
}
\description{
Takes a vector that has been put through scaling and reverses the centering/scaling, if any.
}
